Populate archive and loop packets with PurpleAir measurements.
Incoming archive and loop packets from the user's existing weather station
hardware will now be augmented with measurements made from the PurpleAir
hardware. This will allow those measurements to be potentially archived to
the starndard WeeWX archive database and to posted to RESTFul sources, once
WeeWX is updated to work with these observations.